Implement Scene Camera Integrations

+++ Source/Core/Structures/ERS_STRUCT_SceneCamera/ERS_STRUCT_SceneCamera.h
@@ -26,7 +26,7 @@ struct ERS_STRUCT_SceneCamera {
     glm::vec3 Pos_; /**<Position of the camera object*/
     glm::vec3 Rot_; /**<Rotation of the camera object*/
 
-    std::vector<int> AttachedScriptIndexes_; /**<Indices of scripts that are attached to this camera*/
+    std::vector<long> AttachedScriptIndexes_; /**<Indices of scripts that are attached to this camera*/
     std::string UserDefinedName_; /**<Name that appears in the editor's scene tree*/
 
 };
